Move Firestore implementation into adapter
