--- src/action/firestore/index.ts
@@ -1,4 +1,0 @@
-export * as deck from "@/action/firestore/deck";
-export * as card from "@/action/firestore/card";
-export * as event from "@/action/firestore/event";
-export * as mocked from "@/action/firestore/mocked";
