+++ src/adapters/firestore/event.ts
@@ -1,7 +1,7 @@
 import { onSnapshot, where, collection, query } from "firebase/firestore";
 
-import { mapCardDocument, mapDeckDocument, type CardDocument, type DeckDocument } from "@/action/firestore/dto";
-import { getDb } from "@/firestoreRuntime";
+import { mapCardDocument, mapDeckDocument, type CardDocument, type DeckDocument } from "@/adapters/firestore/dto";
+import { getDb } from "@/adapters/firestore/runtime";
 import type { RemoteChange, RemoteSubscriptionProps } from "@/query/remoteReadContract";
 
 type RemoteEntity = { id: string; updatedAt: number; deletedAt: number | null };
