+++ src/adapters/firestore/rule/rule.spec.ts
