+++ src/firebase.ts
@@ -1,22 +1,11 @@
 import { initializeApp } from "firebase/app";
 import { connectAuthEmulator, getAuth } from "firebase/auth";
 import {
-  connectFirestoreEmulator,
-  initializeFirestore,
-  memoryLocalCache,
-  persistentLocalCache,
-  persistentSingleTabManager,
-  terminate,
-  type Firestore,
-} from "firebase/firestore";
-import {
-  blockFirestoreRuntime,
   getDb,
   getFirestoreInitializationState,
-  initializeFirestoreRuntime,
+  initializeFirestoreAdapter,
   waitForFirestoreInitialization,
-} from "@/firestoreRuntime";
-import { verifyFirestorePersistence } from "@/firestorePersistenceProbe";
+} from "@/adapters/firestore";
 
 const projectId = import.meta.env.VITE_PROJECT_ID;
 const apiKey = import.meta.env.VITE_WEB_API_KEY;
@@ -29,34 +18,7 @@ export const app = initializeApp({
   storageBucket: `${projectId}.appspot.com`,
 });
 export const auth = getAuth(app);
-
-let initializedDb: Firestore | undefined;
-
-try {
-  const localCache = import.meta.env.PROD
-    ? persistentLocalCache({ tabManager: persistentSingleTabManager({}) })
-    : memoryLocalCache();
-  initializedDb = initializeFirestore(app, { localCache });
-  if (import.meta.env.PROD) {
-    const db = initializedDb;
-    void verifyFirestorePersistence(db)
-      .then(() => {
-        initializeFirestoreRuntime(db);
-      })
-      .catch(async (error) => {
-        try {
-          await terminate(db);
-        } catch {
-          // Preserve the persistence failure while terminating best-effort.
-        }
-        blockFirestoreRuntime(error instanceof Error ? error : new Error(String(error)));
-      });
-  } else {
-    initializeFirestoreRuntime(initializedDb);
-  }
-} catch (error) {
-  blockFirestoreRuntime(error instanceof Error ? error : new Error(String(error)));
-}
+initializeFirestoreAdapter(app);
 
 export { getDb, getFirestoreInitializationState, waitForFirestoreInitialization };
 
@@ -64,10 +26,4 @@ if (import.meta.env.MODE === "dev") {
   const host = import.meta.env.VITE_AUTH_HOST;
   const port = import.meta.env.VITE_AUTH_PORT;
   connectAuthEmulator(auth, `http://${host}:${port}`);
-}
-
-if (import.meta.env.DEV && initializedDb) {
-  const host = import.meta.env.VITE_DB_HOST;
-  const port = import.meta.env.VITE_DB_PORT;
-  connectFirestoreEmulator(initializedDb, host, parseInt(port, 10));
 }
